+++ src/basic/components/CartPage/CouponSelector.tsx
@@ -1,0 +1,12 @@
+// CouponSelector 컴포넌트 파일 생성
+// 쿠폰 선택 및 적용 로직을 여기에 작성합니다.
+
+export function CouponSelector({
+  coupons,
+  selectedCoupon,
+  applyCoupon,
+  removeCoupon,
+}: any) {
+  // ...기존 코드 유지...
+  return <div>CouponSelector Component</div>;
+}
