+++ src/basic/components/CartPage/OrderSummary.tsx
@@ -1,0 +1,6 @@
+// OrderSummary 컴포넌트 파일 생성
+// 결제 정보 렌더링 로직을 여기에 작성합니다.
+
+export function OrderSummary({ totals, completeOrder }: any) {
+  return <div>OrderSummary</div>;
+}
