+++ src/basic/components/CartPage/ProductList.tsx
@@ -1,0 +1,12 @@
+// ProductList 컴포넌트 파일 생성
+// 상품 목록 렌더링 로직을 여기에 작성합니다.
+
+export function ProductList({
+  products,
+  searchTerm,
+  addToCart,
+  getRemainingStock,
+  isAdmin,
+}: any) {
+  return <div>ProductList</div>;
+}
